[BCOP-1767] Changed backend to always deal with raw impressions. Changed all instances of impressions to doubles. Fixed numersous instances of precision errors for CPM, Cost, and impressions.
git-tfs-id: [http://devvm01tfs03.telamericamedia.com:8080/tfs/CMWPA]$/Telamerica/Broadcast/OnDeck;C20790

+++ Source/Tam/Maestro/Data/EntityFrameworkMapping.Broadcast/inventory_detail_slot_proposal.cs
@@ -22,7 +22,7 @@ namespace EntityFrameworkMapping.Broadcast
         public int rolled_up_daypart_id { get; set; }
         public decimal slot_cost { get; set; }
         public int spot_length_id { get; set; }
-        public int impressions { get; set; }
+        public double impressions { get; set; }
     
         public virtual proposal_version_detail_quarter_weeks proposal_version_detail_quarter_weeks { get; set; }
     }
